# src/python/dart/web/api/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.dev"
  title: "Dart REST API"
  description: |
    *DRAFT*

    Dart, short for "data mart à la carte", is a self-service data workflow solution that
    facilitates the creation, transformation, and movement of data using data engines,
    primarily in the context of AWS. Put another way, it is a system to facilitate moving
    data in and out of datastores like S3, RDS, Redshift, EMR, DynamoDB, something custom, etc.

    This document describes the Dart REST API using the Swagger 2.0 format. Portions of
    the API are not fully documented due to limitations of the Swagger 2.0 specification
    and the complexity of this API. These APIs will be marked with *NOTE* and will provide
    more details in the descriptive text.
  contact:
    url: https://github.com/RetailMeNotSandbox/dart
  license:
    name: MIT
    url: https://raw.githubusercontent.com/RetailMeNotSandbox/dart/master/LICENSE.txt
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
  /action:
    get:
      operationId: listActions
      summary: "returns the actions matching the filters or all actions if no filters are specified"
      tags:
        - "Action"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
        - $ref: '#/parameters/orderByParam'
        - name: "datastore_id"
          type: "string"
          in: "query"
        - name: "workflow_id"
          type: "string"
          in: "query"
      responses:
        "200":
          description: "returns the actions matching the filters or all actions if no filters are specified."
          schema:
            $ref: '#/definitions/PagedActionsResponse'
  /action/{action_id}:
    get:
      operationId: getAction
      summary: "return the action"
      tags:
        - "Action"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: returns the action
          schema:
            $ref: '#/definitions/ActionResponse'
    put:
      operationId: updateAction
      summary: "update the action"
      tags:
        - "Action"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
        - name: "action"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Action"
      responses:
        "200":
          description: returns the action
          schema:
            $ref: '#/definitions/ActionResponse'
    patch:
      operationId: patchAction
      summary: "patch the action using JSON Patch"
      tags:
        - "Action"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: returns the action
          schema:
            $ref: '#/definitions/ActionResponse'
    delete:
      operationId: deleteAction
      summary: "delete the action"
      tags:
        - "Action"
      parameters:
        - name: "action_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResponse"
...
  /dataset:
    get:
      operationId: listDatasets
      summary: "returns the datasets matching the filters or all datasets if no filters are specified"
      tags:
        - "Dataset"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "returns the datasets matching the filters or all datasets if no filters are specified."
          schema:
            $ref: '#/definitions/PagedDatasetsResponse'
    post:
      operationId: createDataset
      summary: "create a dataset"
      tags:
        - "Dataset"
      parameters:
        - name: "dataset"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Dataset"
      responses:
        "200":
          description: returns the created Dataset
          schema:
            $ref: '#/definitions/DatasetResponse'
  /dataset/guess:
    get:
      operationId: guessDataset
      summary: "best guess of the dataset"
      tags:
        - "Dataset"
      parameters:
        - name: s3_path
          in: query
          type: string
          required: true
        - name: max_lines
          in: query
          type: integer
          required: true
      responses:
        "200":
          description: returns the dataset
          schema:
            $ref: '#/definitions/DatasetResponse'
  /dataset/{dataset_id}:
    get:
      operationId: getDataset
      summary: "get a dataset"
      tags:
        - "Dataset"
      parameters:
        - name: "dataset_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: returns the dataset
          schema:
            $ref: '#/definitions/DatasetResponse'
    put:
      operationId: updateDataset
      summary: "update a dataset"
      tags:
        - "Dataset"
      parameters:
        - name: "dataset_id"
          type: "string"
          in: "path"
          required: true
        - name: "dataset"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Dataset"
      responses:
        "200":
          description: returns the dataset
          schema:
            $ref: '#/definitions/DatasetResponse'
    delete:
      operationId: deleteDataset
      summary: "delete a dataset"
      tags:
        - "Dataset"
      parameters:
        - name: "dataset_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResponse"
...
  /datastore:
    post:
      operationId: createDatastore
      summary: "create a datastore"
      tags:
        - "Datastore"
      parameters:
        - name: "datastore"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Datastore"
      responses:
        "200":
          description: returns the created datastore
          schema:
            $ref: '#/definitions/DatastoreResponse'
    get:
      operationId: listDatastores
      summary: "list datastores"
      tags:
        - "Datastore"
      parameters:
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/offsetParam'
        - $ref: '#/parameters/filtersParam'
      responses:
        "200":
          description: "returns the datastores matching the filters or all datastores if no filters are specified."
          schema:
            $ref: '#/definitions/PagedDatastoresResponse'
  /datastore/{datastore_id}:
    get:
      operationId: getDatastore
      summary: "return the datastore"
      tags:
        - "Datastore"
      parameters:
        - name: "datastore_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: returns the datastore
          schema:
            $ref: '#/definitions/DatastoreResponse'
    put:
      operationId: updateDatastore
      summary: "update the datastore"
      tags:
        - "Datastore"
      parameters:
        - name: "datastore_id"
          type: "string"
          in: "path"
          required: true
        - name: "datastore"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Datastore"
      responses:
        "200":
          description: returns the datastore
          schema:
            $ref: '#/definitions/DatastoreResponse'
        "400":
          description: "datastore is not in a valid state"
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      operationId: patchDatastore
      summary: "patch the datastore using JSON Patch"
      tags:
        - "Datastore"
      parameters:
        - name: "datastore_id"
          type: "string"
          in: "path"
          required: true
        - $ref: '#/parameters/patchParam'
      responses:
        "200":
          description: returns the datastore
          schema:
            $ref: '#/definitions/DatastoreResponse'
        "400":
          description: "datastore is not in a valid state"
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteDatastore
      summary: "delete the datastore"
      tags:
        - "Datastore"
      parameters:
        - name: "datastore_id"
          type: "string"
          in: "path"
          required: true
      responses:
        "200":
          description: "returns a success message"
          schema:
            $ref: "#/definitions/OKResponse"
  /datastore/{datastore_id}/action:
    post:
      operationId: createDatastoreActions
      summary: "create actions for a datastore"
      tags:
        - "Datastore"
        - "Action"
      parameters:
        - name: "datastore_id"
          type: "string"
          in: "path"
          required: true
        - name: "actions"
          in: "body"
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/Action"
          description: |
            ~~~~
            NOTE: This parameter is specified as an array because Swagger 2.0 does
            not support the oneOf construct. This API also supports passing a single
            Action object.
            ~~~~
      responses:
        "200":
          description: "the created actions"
          schema:
            $ref: '#/definitions/ActionsResponse'
...
  /workflow/{workflow_id}/action:
    post:
      operationId: createWorkflowActions
      summary: "create actions for a workflow"
      tags:
        - "Workflow"
        - "Action"
      parameters:
        - name: "workflow_id"
          description: "the workflow id"
          type: "string"
          in: "path"
          required: true
        - name: "actions"
          in: "body"
          required: true
          schema:
            type: array
            items:
              $ref: "#/definitions/Action"
          description: |
            ~~~~
            NOTE: This parameter is specified as an array because Swagger 2.0 does
            not support the oneOf construct. This API also supports passing a single
            Action object.
            ~~~~
      responses:
        "200":
          description: "the created actions"
          schema:
            $ref: '#/definitions/ActionsResponse'
...
parameters:
  limitParam:
    name: "limit"
    type: "integer"
    in: "query"
    description: |
      Specify the maximum number of records that should be returned by this request.
  offsetParam:
    name: "offset"
    type: "integer"
    in: "query"
    description: |
      Specify the starting index for the next chunk of results to be returned by this request.
  filtersParam:
    name: filters
    in: query
    type: string
    description: |
      ~~~~
      NOTE: This parameter is specified as a string because Swagger 2.0 does
      not support passing object or array types as query string parameters.
      ~~~~
      To use, pass a URL encoded JSON array of Filter objects (as defined in
      the definitions section).
  orderByParam:
    name: "order_by"
    type: string
    in: query
    description: |
      ~~~~
      NOTE: This parameter is specified as a string because Swagger 2.0 does
      not support passing object or array types as query string parameters.
      ~~~~
      To use, pass a URL encoded JSON array of OrderBy objects (as defined in
      the definitions section).
  patchParam:
    name: "patch"
    description: "a JSON Patch object"
    in: "body"
    required: true
    schema:
      $ref: '#/definitions/JSONPatch'
#
# Definitions are auto-generated by dart.tool.export_swagger_definitions
# Do not manually edit!
#
definitions:
  Action:
    additionalProperties: false
    properties:
      created:
        type: string
        x-nullable: true
      data:
        $ref: '#/definitions/ActionData'
      id:
        type: string
        x-nullable: true
      updated:
        type: string
        x-nullable: true
      version_id:
        type: integer
        x-nullable: true
    required:
    - data
    type: object
...
  ActionData:
    additionalProperties: false
    properties:
      action_type_name:
        type: string
      args:
        type: object
        x-nullable: true
      avg_runtime:
        type: string
        x-nullable: true
      completed_runs:
        default: 0
        type: integer
      datastore_id:
        type: string
        x-nullable: true
      ecs_task_arn:
        type: string
        x-nullable: true
      end_time:
        type: string
        x-nullable: true
      engine_name:
        type: string
      error_message:
        type: string
        x-nullable: true
        x-schema-form:
          type: textarea
      extra_data:
        type: object
        x-nullable: true
      first_in_workflow:
        default: false
        type: boolean
        x-nullable: true
      last_in_workflow:
        default: false
        type: boolean
        x-nullable: true
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      on_failure:
        default: DEACTIVATE
        description: applies to the workflow if this is a workflow action template,
          otherwise the datastore
        enum:
        - DEACTIVATE
        - CONTINUE
        type: string
      on_failure_email:
        default: []
        items:
          pattern: ^\S+@\S+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''email '' + $index }}'
          type: tabarray
      on_success_email:
        default: []
        items:
          pattern: ^\S+@\S+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''email '' + $index }}'
          type: tabarray
      order_idx:
        minimum: 0.0
        type: number
        x-nullable: true
      progress:
        type: number
        x-nullable: true
      queued_time:
        type: string
        x-nullable: true
      start_time:
        type: string
        x-nullable: true
      state:
        default: HAS_NEVER_RUN
        enum:
        - HAS_NEVER_RUN
        - QUEUED
        - PENDING
        - RUNNING
        - FINISHING
        - COMPLETED
        - FAILED
        - SKIPPED
        - TEMPLATE
        type: string
      tags:
        default: []
        items:
          maxLength: 40
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      user_id:
        default: unknown
        maxLength: 64
        pattern: ^[@.a-zA-Z0-9_-]+$
        type: string
      workflow_action_id:
        type: string
        x-nullable: true
      workflow_id:
        type: string
        x-nullable: true
      workflow_instance_id:
        type: string
        x-nullable: true
    required:
    - action_type_name
    - engine_name
    - name
    - args
    - user_id
    type: object
  ActionResponse:
    properties:
      results:
        $ref: '#/definitions/Action'
    type: object
...
  ActionsResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/Action'
        type: array
    type: object
...
  Dataset:
    additionalProperties: false
    properties:
      created:
        type: string
        x-nullable: true
      data:
        $ref: '#/definitions/DatasetData'
      id:
        type: string
        x-nullable: true
      updated:
        type: string
        x-nullable: true
      version_id:
        type: integer
        x-nullable: true
    required:
    - data
    type: object
  DatasetData:
    additionalProperties: false
    properties:
      batch_merge_sort_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+ (ASC|DESC)$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''upsert_sort_key '' + $index }}'
          type: tabarray
      columns:
        items:
          $ref: '#/definitions/Column'
        minItems: 1
        type: array
        x-schema-form:
          title: '{{ value.name || ''column '' + $index }}'
          type: tabarray
      compression:
        enum:
        - NONE
        - BZ2
        - GZIP
        - SNAPPY
        type: string
      data_format:
        $ref: '#/definitions/DataFormat'
      description:
        type: string
        x-nullable: true
      distribution_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''distribution_key '' + $index }}'
          type: tabarray
      hive_compatible_partition_folders:
        default: false
        type: boolean
        x-nullable: true
      load_type:
        enum:
        - INSERT
        - MERGE
        - RELOAD_ALL
        - RELOAD_LAST
        type: string
      location:
        pattern: ^s3://.+$
        type: string
      merge_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''merge_key '' + $index }}'
          type: tabarray
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      partitions:
        items:
          $ref: '#/definitions/Column'
        minItems: 0
        type: array
        x-nullable: true
        x-schema-form:
          title: '{{ value.name || ''column '' + $index }}'
          type: tabarray
      primary_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''primary_key '' + $index }}'
          type: tabarray
      sort_keys:
        default: []
        items:
          maxLength: 127
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        type: array
        x-schema-form:
          title: '{{ value || ''sort_key '' + $index }}'
          type: tabarray
      table_name:
        minLength: 1
        pattern: ^[a-zA-Z0-9_]+$
        type: string
      tags:
        default: []
        items:
          maxLength: 40
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      user_id:
        default: unknown
        maxLength: 64
        pattern: ^[@.a-zA-Z0-9_-]+$
        type: string
    required:
    - name
    - table_name
    - location
    - load_type
    - data_format
    - columns
    - compression
    - user_id
    type: object
  DatasetResponse:
    properties:
      results:
        $ref: '#/definitions/Dataset'
    type: object
  Datastore:
    additionalProperties: false
    properties:
      created:
        type: string
        x-nullable: true
      data:
        $ref: '#/definitions/DatastoreData'
      id:
        type: string
        x-nullable: true
      updated:
        type: string
        x-nullable: true
      version_id:
        type: integer
        x-nullable: true
    required:
    - data
    type: object
  DatastoreData:
    additionalProperties: false
    properties:
      args:
        type: object
        x-nullable: true
      concurrency:
        default: 1
        maximum: 10
        minimum: 1
        type: integer
      connection_url:
        type: string
        x-nullable: true
      engine_name:
        pattern: ^[a-zA-Z0-9_]+$
        type: string
      extra_data:
        type: object
        x-nullable: true
      host:
        type: string
        x-nullable: true
      name:
        maxLength: 50
        pattern: ^[a-zA-Z0-9_-]+$
        type: string
      port:
        type: integer
        x-nullable: true
      s3_artifacts_path:
        type: string
        x-nullable: true
      s3_logs_path:
        type: string
        x-nullable: true
      state:
        default: INACTIVE
        enum:
        - INACTIVE
        - ACTIVE
        - DONE
        - TEMPLATE
        type: string
      tags:
        default: []
        items:
          maxLength: 40
          pattern: ^[a-zA-Z0-9_]+$
          type: string
        minItems: 0
        type: array
        x-schema-form:
          title: '{{ value || ''tag '' + $index }}'
          type: tabarray
      user_id:
        default: unknown
        maxLength: 64
        pattern: ^[@.a-zA-Z0-9_-]+$
        type: string
      workflow_datastore_id:
        type: string
        x-nullable: true
    required:
    - name
    - engine_name
    - args
    - user_id
    type: object
  DatastoreResponse:
    properties:
      results:
        $ref: '#/definitions/Datastore'
    type: object
...
  ErrorResponse:
    properties:
      error_message:
        type: string
      results:
        enum:
        - ERROR
        type: string
    type: object
...
  JSONPatch:
    description: JSON Patch object as defined by http://json.schemastore.org/json-patch
    type: object
...
  OKResponse:
    properties:
      results:
        enum:
        - OK
        type: string
    type: object
...
  PagedActionsResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/Action'
        type: array
      total:
        type: integer
    type: object
  PagedDatasetsResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/Dataset'
        type: array
      total:
        type: integer
    type: object
  PagedDatastoresResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/Datastore'
        type: array
      total:
        type: integer
    type: object
...